Fixed issue #9 in mwielgoszewski/jython-burp-api: Missing ps1 object
For some reason, Burp is not loading the jython package correctly because it does not find the ps1 or ps2 references in the sys module. From the command line, we can see jython's `sys` has both. For now, the easy fix is to replace the references with their hard coded strings.